feat(frontend): Implement authentication helper hook for token refresh

+++ client/src/hooks/useHandleListings.tsx
@@ -1,24 +1,36 @@
 import { useAppDispatch } from "../redux/hooks";
 import { setUser } from "../redux/features/userSlice";
 import { toast } from "react-toastify";
+import useAuthentication from "./useAuthentication";
 function useHandleListings() {
+  const { refreshToken } = useAuthentication();
   const dispatch = useAppDispatch();
   const handleDeleteListing = async (id: string) => {
-    try {
-      const res = await fetch(`/api/property/${id}`, {
-        method: "DELETE",
-      });
-      const data = await res.json();
-      if (!data.success) {
-        toast.error("something went wrong");
+    refreshToken().then(async (result) => {
+      if (!result?.isAuthenticated) {
+        return;
       }
-      toast.success("listing has been deleted");
-      dispatch(setUser(data.user));
-    } catch (error) {
-      if (error instanceof Error) {
-        toast.error("something went wrong");
+      try {
+        const res = await fetch(`/api/property/${id}`, {
+          credentials: "include",
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            authorization: `Bearer ${result.accessToken}`,
+          },
+        });
+        const data = await res.json();
+        if (!data.success) {
+          toast.error("something went wrong");
+        }
+        toast.success("listing has been deleted");
+        dispatch(setUser(data.user));
+      } catch (error) {
+        if (error instanceof Error) {
+          toast.error("something went wrong");
+        }
       }
-    }
+    });
   };
   return { handleDeleteListing };
 }
